requirements.txt
requests
python-vlc